requirements.txt
streamlit==1.40.1
pandas==1.5.3
nltk==3.8.1
pyspellchecker==0.8.1
difflib-json==0.1.1
PyPDF2==3.0.1